HB-3431] Reference adapter
A "reference" adapter used for showcasing as well as testing the Helium mediation contract. Internal use only. Accompanied by a dummy SDK with a trivial implementation.

One piece I've glossed over is the fullscreen ad callback chain (from the dummy SDK back to the reference adapter). As the ad is shown in a new Activity, we lose track of the ad and its events, so for the sake of simplicity I'm triggering all partner callbacks together before launching the Activity. Technically I could build the callback chain via traditional listener, or EventBus, or BroadcastReceiver, but I'm already using lambdas for callbacks (which is nice!), so I didn't really want to introduce another callback system.

This should be code complete but will need to be tested (contingent upon HB-3903). Some of the XML nav logic might need additional work as a result.